Your commit message

+++ src/controllers/admin.ts
@@ -4,7 +4,7 @@ import { Lecture } from "../models/Lecture.js";
 import { rm } from "fs";
 import { promisify } from "util";
 import fs from "fs";
-import { IUser, User } from "../models/User.js";
+import { User } from "../models/User.js";
 export const createCourse = TryCatch(async (req: any, res: any) => {
   const { title, description, category, createdBy, duration, price } = req.body;
 
@@ -53,9 +53,10 @@ export const addLectures = TryCatch(async (req: any, res: any) => {
 export const deleteLecture = TryCatch(async (req: any, res: any) => {
   const lecture = await Lecture.findById(req.params.id);
 
-  rm(lecture.video, () => {
-    console.log("Video deleted");
-  });
+  if (!lecture) return res.status(404).json({ message: "Lecture not found" });
+rm(lecture.video, () => { console.log("Video deleted"); });
+
+
 
   await lecture.deleteOne();
 
@@ -67,7 +68,14 @@ const unlinkAsync = promisify(fs.unlink);
 export const deleteCourse = TryCatch(async (req: any, res: any) => {
   const course = await Courses.findById(req.params.id);
 
+  
+
+  if (!course) {
+    return res.status(404).json({ message: "Course not found" });
+  }
+  
   const lectures = await Lecture.find({ course: course._id });
+  
 
   await Promise.all(
     lectures.map(async (lecture) => {
@@ -76,14 +84,21 @@ export const deleteCourse = TryCatch(async (req: any, res: any) => {
     })
   );
 
+ 
+
+  if (!course) {
+    return res.status(404).json({ message: "Course not found" });
+  }
+  
   rm(course.image, () => {
     console.log("image deleted");
   });
+  
 
   await Lecture.find({ course: req.params.id }).deleteMany();
 
   await course.deleteOne();
-
+ 
   await User.updateMany({}, { $pull: { subscription: req.params.id } });
 
   res.json({
@@ -120,23 +135,28 @@ export const updateRole = TryCatch(async (req: any, res: any) => {
     return res.status(403).json({
       message: "This endpoint is assign to superadmin",
     });
-  const user = await User.findById(req.params.id);
+    const user = await User.findById(req.params.id);
 
-  if (user.role === "user") {
-    user.role = "admin";
-    await user.save();
-
-    return res.status(200).json({
-      message: "Role updated to admin",
-    });
-  }
-
-  if (user.role === "admin") {
-    user.role = "user";
-    await user.save();
-
-    return res.status(200).json({
-      message: "Role updated",
-    });
-  }
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
+    
+    if (user.role === "user") {
+      user.role = "admin";
+      await user.save();
+    
+      return res.status(200).json({
+        message: "Role updated to admin",
+      });
+    }
+    
+    if (user.role === "admin") {
+      user.role = "user";
+      await user.save();
+    
+      return res.status(200).json({
+        message: "Role updated",
+      });
+    }
+    
 });
